requirements.txt
anyio==3.6.2
certifi==2022.12.7
clarifai-grpc ==9.2.0
cffi==1.15.1
cryptography==39.0.2
decorator==5.1.1
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
pycparser==2.21
python-telegram-bot==12.7
rfc3986==1.5.0
sniffio==1.3.0
socksio==1.0.0
tornado==6.2
emoji==2.2.0
dnspython==2.3.0 
pymongo==4.3.3
pytz==2023.3